finalisation de l'exo 1 du TP4
Implémentation complète de la fenêtre modale pour ajouter un jeu
