+++ GUI/GUI.csproj
@@ -47,11 +47,11 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="core\Catalogue.cs" />
-    <Compile Include="Form1.cs">
+    <Compile Include="MainForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Form1.Designer.cs">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <Compile Include="MainForm.Designer.cs">
+      <DependentUpon>MainForm.cs</DependentUpon>
     </Compile>
     <Compile Include="core\Genre.cs" />
     <Compile Include="core\IFieldComparable.cs" />
@@ -67,8 +67,8 @@
       <DependentUpon>SaisieJeuDlg.cs</DependentUpon>
     </Compile>
     <Compile Include="utilities\QuickSort.cs" />
-    <EmbeddedResource Include="Form1.resx">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <EmbeddedResource Include="MainForm.resx">
+      <DependentUpon>MainForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
